+++ src/main/java/com/seudev/util/message/DefaultMessageInterpolator.java
@@ -1,4 +1,4 @@
-package com.infinityrefactoring.util.message;
+package com.seudev.util.message;
 
 import static java.util.logging.Level.FINEST;
 
@@ -13,46 +13,46 @@ import javax.enterprise.context.RequestScoped;
 import javax.enterprise.inject.Default;
 import javax.inject.Inject;
 
-import com.infinityrefactoring.util.text.Expression;
-import com.infinityrefactoring.util.text.ExpressionDefinitions.DollarCurlyBracket;
-import com.infinityrefactoring.util.text.ExpressionException;
+import com.seudev.util.text.Expression;
+import com.seudev.util.text.ExpressionDefinitions.DollarCurlyBracket;
+import com.seudev.util.text.ExpressionException;
 
 @Default
 @RequestScoped
 public class DefaultMessageInterpolator implements MessageInterpolator {
-    
+
     @Inject
     private Logger logger;
-    
+
     @Inject
     private MessageProvider messageProvider;
-    
+
     @Inject
     private DollarCurlyBracket dollarCurlyBracket;
-    
+
     @Inject
     private ELProcessor elProcessor;
-    
+
     @Inject
     private List<Locale> locales;
-    
+
     @Override
     public MessageInterpolator add(String key, Object value) {
         elProcessor.defineBean(key, value);
         return this;
     }
-    
+
     @Override
     public String get(String key, boolean required) {
         return get(key, locales, required);
     }
-    
+
     @Override
     public String get(String key, List<Locale> locales, boolean required) {
         if (key == null) {
             throw new IllegalArgumentException("The key must be not null");
         }
-        
+
         String template = null;
         for (Locale locale : locales) {
             template = messageProvider.getMessages(locale).get(key);
@@ -60,7 +60,7 @@ public class DefaultMessageInterpolator implements MessageInterpolator {
                 break;
             }
         }
-        
+
         if (template == null) {
             if (required) {
                 throw new IllegalArgumentException("Not found message for key: " + key);
@@ -69,7 +69,7 @@ public class DefaultMessageInterpolator implements MessageInterpolator {
         }
         return interpolate(key, template);
     }
-    
+
     private String interpolate(String key, String template) {
         try {
             SortedMap<Expression, SortedSet<Integer>> expressions = dollarCurlyBracket.findAll(template, 0);
@@ -82,5 +82,5 @@ public class DefaultMessageInterpolator implements MessageInterpolator {
             throw new ExpressionException("Cannot be interpolate the message: " + key, ex);
         }
     }
-    
+
 }
