+++ src/main/java/com/seudev/util/message/PropertyMessageSource.java
@@ -1,4 +1,4 @@
-package com.infinityrefactoring.util.message;
+package com.seudev.util.message;
 
 import static java.util.logging.Level.FINEST;
