+++ src/main/java/com/seudev/util/message/package-info.java
@@ -1,4 +1,4 @@
 /**
  * @author Thomás Sousa Silva (ThomasSousa96)
  */
-package com.infinityrefactoring.util.message;
+package com.seudev.util.message;
